FIX: Refactor after rebase
